move configmap functions to their own file
